+++ core/rollout/request.go
@@ -25,7 +25,7 @@ type Request struct {
 type RequestBundle struct {
 	Version         string `json:"version"`
 	ReleaseSequence int    `json:"release_sequence"`
-	Channel         string `json:"channel"`
+	Channel         string `json:"channel,omitempty"`
 	ArtifactDigest  string `json:"artifact_digest"`
 	ManifestDigest  string `json:"manifest_digest"`
 }